+++ ambit2-apps/ambit2-www/src/main/java/ambit2/rest/AmbitApplication.java
@@ -500,9 +500,9 @@ public class AmbitApplication extends AmbitFreeMarkerApplication<Object> {
 
 			if (isDBAAEnabled()) {
 				String secret = getProperty(AMBITConfig.secret.name(), configProperties);
-				long sessionLength = 1000 * 60 * 45L; // 45 min in milliseconds
+				int sessionLength = 1000 * 60 * 45; // 45 min in milliseconds
 				try {
-					sessionLength = Long.parseLong(getProperty(AMBITConfig.sessiontimeout.name(), configProperties));
+					sessionLength = Integer.parseInt(getProperty(AMBITConfig.sessiontimeout.name(), configProperties));
 				} catch (Exception x) {
 				}
